+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/cursors/AutoContinuingCursor.java
@@ -20,9 +20,9 @@
 
 package com.apple.foundationdb.record.cursors;
 
-import com.apple.foundationdb.FDBException;
 import com.apple.foundationdb.annotation.API;
 import com.apple.foundationdb.async.AsyncUtil;
+import com.apple.foundationdb.async.MoreAsyncUtil;
 import com.apple.foundationdb.record.RecordCursor;
 import com.apple.foundationdb.record.RecordCursorResult;
 import com.apple.foundationdb.record.RecordCursorVisitor;
@@ -74,8 +74,6 @@ import java.util.function.BiFunction;
 public class AutoContinuingCursor<T> implements RecordCursor<T> {
     private static final Logger LOGGER = LoggerFactory.getLogger(AutoContinuingCursor.class);
 
-    private static final int MAX_RETRY_ON_EXCEPTION_ATTEMPTS = 3;
-
     @Nonnull
     private final FDBDatabaseRunner runner;
     @Nonnull
@@ -89,7 +87,7 @@ public class AutoContinuingCursor<T> implements RecordCursor<T> {
     @Nullable
     private RecordCursorResult<T> lastResult;
 
-    private final int maxAttemptsOnRetryableExceptions;
+    private final int maxRetriesOnRetriableException;
 
     /**
      * Creates a new {@link AutoContinuingCursor}.
@@ -105,16 +103,16 @@ public class AutoContinuingCursor<T> implements RecordCursor<T> {
      * Creates a new {@link AutoContinuingCursor}.
      * @param runner the runner from which it can open new contexts
      * @param nextCursorGenerator the method which can generate the underlying cursor given a record context and a continuation
-     * @param maxAttemptsOnRetryableExceptions maximum number of consecutive times retryable exceptions, such as
+     * @param maxRetriesOnRetriableException maximum number of consecutive times retryable exceptions, such as
      *   {@link com.apple.foundationdb.FDBError#TRANSACTION_TOO_OLD}, will be caught and a the cursor automatically
      *   continued
      */
     public AutoContinuingCursor(@Nonnull FDBDatabaseRunner runner,
                                 @Nonnull BiFunction<FDBRecordContext, byte[], RecordCursor<T>> nextCursorGenerator,
-                                int maxAttemptsOnRetryableExceptions) {
+                                int maxRetriesOnRetriableException) {
         this.runner = runner;
         this.nextCursorGenerator = nextCursorGenerator;
-        this.maxAttemptsOnRetryableExceptions = maxAttemptsOnRetryableExceptions;
+        this.maxRetriesOnRetriableException = maxRetriesOnRetriableException;
     }
 
     @Nonnull
@@ -138,33 +136,13 @@ public class AutoContinuingCursor<T> implements RecordCursor<T> {
             openContextAndGenerateCursor(null);
         }
 
-        return currentCursor.onNext().exceptionally(exception -> {
-            if (!isRetryable(exception) || attempt >= maxAttemptsOnRetryableExceptions) {
-                if (exception instanceof RuntimeException) {
-                    throw (RuntimeException)exception;
-                }
+        return MoreAsyncUtil.handleOnException(() -> currentCursor.onNext(), exception -> {
+            if (!FDBExceptions.isRetriable(exception) || attempt >= maxRetriesOnRetriableException) {
                 throw FDBExceptions.wrapException(exception);
             }
             openContextAndGenerateCursor(lastResult == null ? null : lastResult.getContinuation().toBytes());
-
-            // Null signals the thenCompose(), below, to compose another attempt at onNext()
-            return null;
-        }).thenCompose(result -> {
-            if (result == null) {
-                return onNextWithRetry(attempt + 1);
-            }
-            return CompletableFuture.completedFuture(result);
+            return onNextWithRetry(attempt + 1);
         });
-    }
-
-    private boolean isRetryable(Throwable e) {
-        while (e != null) {
-            if (e instanceof FDBException && ((FDBException) e).isRetryable()) {
-                return true;
-            }
-            e = e.getCause();
-        }
-        return false;
     }
 
     @Nonnull