+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBExceptions.java
@@ -155,10 +155,6 @@ public class FDBExceptions {
     }
 
     public static RuntimeException wrapException(@Nonnull Throwable ex) {
-        if (ex instanceof RecordCoreException) {
-            return (RecordCoreException) ex;
-        }
-
         //transfer any logging details into mapped exception
         Object[] logInfo;
         if (ex instanceof LoggableKeysAndValues) {