+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/keyspace/LocatableResolverTest.java
@@ -1166,7 +1166,7 @@ public abstract class LocatableResolverTest extends FDBTestBase {
                             .forEach(validated -> { }).thenCompose(vignore -> context.commitAsync()));
         }
 
-        validate(globalScope, invalidEntries);
+        validate(globalScope, Collections.emptySet());
     }
 
     private void validate(LocatableResolver locatableResolver, Set<ResolverValidator.ValidatedEntry> expectedBadEntries) {